--- a/.github/config/extensions/ducklake.cmake
+++ b/.github/config/extensions/ducklake.cmake
@@ -2,4 +2,5 @@ duckdb_extension_load(ducklake
     DONT_LINK
     GIT_URL https://github.com/duckdb/ducklake
     GIT_TAG d2392c36f33151cf5cdd7d006375b0b669bd44ac
+    APPLY_PATCHES
 )