--- a/.github/patches/extensions/mysql_scanner/fix.patch
+++ b/.github/patches/extensions/mysql_scanner/fix.patch
@@ -218,7 +218,7 @@ index 587cdf2..ce95653 100644
  
  //===--------------------------------------------------------------------===//
 diff --git a/src/storage/mysql_insert.cpp b/src/storage/mysql_insert.cpp
-index 121d85a..0d84183 100644
+index 121d85a..395111b 100644
 --- a/src/storage/mysql_insert.cpp
 +++ b/src/storage/mysql_insert.cpp
 @@ -14,14 +14,14 @@
@@ -239,4 +239,13 @@ index 121d85a..0d84183 100644
 +    : PhysicalOperator(physical_plan, PhysicalOperatorType::EXTENSION, op.types, 1), table(nullptr), schema(&schema),
        info(std::move(info)) {
  }
+ 
+@@ -307,7 +307,7 @@ PhysicalOperator &MySQLCatalog::PlanInsert(ClientContext &context, PhysicalPlanG
+ 	if (op.return_chunk) {
+ 		throw BinderException("RETURNING clause not yet supported for insertion into MySQL table");
+ 	}
+-	if (op.action_type != OnConflictAction::THROW) {
++	if (op.on_conflict_info.action_type != OnConflictAction::THROW) {
+ 		throw BinderException("ON CONFLICT clause not yet supported for insertion into MySQL table");
+ 	}