--- a/.github/patches/extensions/sqlite_scanner/fix.patch
+++ b/.github/patches/extensions/sqlite_scanner/fix.patch
@@ -196,7 +196,7 @@ index 440db80..e076502 100644
  
  //===--------------------------------------------------------------------===//
 diff --git a/src/storage/sqlite_insert.cpp b/src/storage/sqlite_insert.cpp
-index 9cd4786..fbca2ac 100644
+index 9cd4786..0339e6b 100644
 --- a/src/storage/sqlite_insert.cpp
 +++ b/src/storage/sqlite_insert.cpp
 @@ -13,14 +13,14 @@
@@ -217,6 +217,15 @@ index 9cd4786..fbca2ac 100644
 +    : PhysicalOperator(physical_plan, PhysicalOperatorType::EXTENSION, op.types, 1), table(nullptr), schema(&schema),
        info(std::move(info)) {
  }
+ 
+@@ -184,7 +184,7 @@ PhysicalOperator &SQLiteCatalog::PlanInsert(ClientContext &context, PhysicalPlan
+ 	if (op.return_chunk) {
+ 		throw BinderException("RETURNING clause not yet supported for insertion into SQLite table");
+ 	}
+-	if (op.action_type != OnConflictAction::THROW) {
++	if (op.on_conflict_info.action_type != OnConflictAction::THROW) {
+ 		throw BinderException("ON CONFLICT clause not yet supported for insertion into SQLite table");
+ 	}
  
 diff --git a/src/storage/sqlite_update.cpp b/src/storage/sqlite_update.cpp
 index b550549..6c071d7 100644