--- a/src/execution/physical_plan/plan_insert.cpp
+++ b/src/execution/physical_plan/plan_insert.cpp
@@ -108,11 +108,11 @@ PhysicalOperator &DuckCatalog::PlanInsert(ClientContext &context, PhysicalPlanGe
 		parallel_streaming_insert = false;
 		use_batch_index = false;
 	}
-	if (op.action_type != OnConflictAction::THROW) {
+	if (op.on_conflict_info.action_type != OnConflictAction::THROW) {
 		// We don't support ON CONFLICT clause in batch insertion operation currently
 		use_batch_index = false;
 	}
-	if (op.action_type == OnConflictAction::UPDATE) {
+	if (op.on_conflict_info.action_type == OnConflictAction::UPDATE) {
 		// When we potentially need to perform updates, we have to check that row is not updated twice
 		// that currently needs to be done for every chunk, which would add a huge bottleneck to parallelized insertion
 		parallel_streaming_insert = false;
@@ -128,11 +128,11 @@ PhysicalOperator &DuckCatalog::PlanInsert(ClientContext &context, PhysicalPlanGe
 	}
 
 	auto &insert = planner.Make<PhysicalInsert>(
-	    op.types, op.table, std::move(op.bound_constraints), std::move(op.expressions), std::move(op.set_columns),
-	    std::move(op.set_types), op.estimated_cardinality, op.return_chunk,
-	    parallel_streaming_insert && num_threads > 1, op.action_type, std::move(op.on_conflict_condition),
-	    std::move(op.do_update_condition), std::move(op.on_conflict_filter), std::move(op.columns_to_fetch),
-	    op.update_is_del_and_insert);
+	    op.types, op.table, std::move(op.bound_constraints), std::move(op.expressions), std::move(op.on_conflict_info.set_columns),
+	    std::move(op.on_conflict_info.set_types), op.estimated_cardinality, op.return_chunk,
+	    parallel_streaming_insert && num_threads > 1, op.on_conflict_info.action_type, std::move(op.on_conflict_info.on_conflict_condition),
+	    std::move(op.on_conflict_info.do_update_condition), std::move(op.on_conflict_info.on_conflict_filter), std::move(op.on_conflict_info.columns_to_fetch),
+	    op.on_conflict_info.update_is_del_and_insert);
 	insert.children.push_back(*plan);
 	return insert;
 }