--- a/src/include/duckdb/planner/bound_tokens.hpp
+++ b/src/include/duckdb/planner/bound_tokens.hpp
@@ -34,6 +34,7 @@ class BoundConjunctionExpression;
 class BoundConstantExpression;
 class BoundDefaultExpression;
 class BoundFunctionExpression;
+class BoundLambdaRefExpression;
 class BoundOperatorExpression;
 class BoundParameterExpression;
 class BoundReferenceExpression;