--- a/src/include/duckdb/planner/operator/logical_insert.hpp
+++ b/src/include/duckdb/planner/operator/logical_insert.hpp
@@ -17,28 +17,8 @@ class TableCatalogEntry;
 
 class Index;
 
-//! LogicalInsert represents an insertion of data into a base table
-class LogicalInsert : public LogicalOperator {
-public:
-	static constexpr const LogicalOperatorType TYPE = LogicalOperatorType::LOGICAL_INSERT;
-
-public:
-	LogicalInsert(TableCatalogEntry &table, idx_t table_index);
-
-	vector<vector<unique_ptr<Expression>>> insert_values;
-	//! The insertion map ([table_index -> index in result, or DConstants::INVALID_INDEX if not specified])
-	physical_index_vector_t<idx_t> column_index_map;
-	//! The expected types for the INSERT statement (obtained from the column types)
-	vector<LogicalType> expected_types;
-	//! The base table to insert into
-	TableCatalogEntry &table;
-	idx_t table_index;
-	//! if returning option is used, return actual chunk to projection
-	bool return_chunk;
-	//! The default statements used by the table
-	vector<unique_ptr<Expression>> bound_defaults;
-	//! The constraints used by the table
-	vector<unique_ptr<BoundConstraint>> bound_constraints;
+struct BoundOnConflictInfo {
+	BoundOnConflictInfo();
 
 	//! Which action to take on conflict
 	OnConflictAction action_type;
@@ -62,6 +42,32 @@ public:
 	vector<column_t> source_columns;
 	//! True, if the INSERT OR REPLACE requires delete + insert.
 	bool update_is_del_and_insert;
+};
+
+//! LogicalInsert represents an insertion of data into a base table
+class LogicalInsert : public LogicalOperator {
+public:
+	static constexpr const LogicalOperatorType TYPE = LogicalOperatorType::LOGICAL_INSERT;
+
+public:
+	LogicalInsert(TableCatalogEntry &table, idx_t table_index);
+
+	vector<vector<unique_ptr<Expression>>> insert_values;
+	//! The insertion map ([table_index -> index in result, or DConstants::INVALID_INDEX if not specified])
+	physical_index_vector_t<idx_t> column_index_map;
+	//! The expected types for the INSERT statement (obtained from the column types)
+	vector<LogicalType> expected_types;
+	//! The base table to insert into
+	TableCatalogEntry &table;
+	idx_t table_index;
+	//! if returning option is used, return actual chunk to projection
+	bool return_chunk;
+	//! The default statements used by the table
+	vector<unique_ptr<Expression>> bound_defaults;
+	//! The constraints used by the table
+	vector<unique_ptr<BoundConstraint>> bound_constraints;
+
+	BoundOnConflictInfo on_conflict_info;
 
 public:
 	void Serialize(Serializer &serializer) const override;