… benchmark/tpch/upsert/bulk_upsert_empty_table.benchmark
… benchmark/tpch/upsert/bulk_upsert_full_table.benchmark
… benchmark/tpch/upsert/tiny_insert_full_table.benchmark
… src/include/duckdb/storage/serialization/logical_operator.json
… src/planner/binder/statement/bind_merge_into.cpp
… src/planner/expression_binder/projection_binder.cpp
… src/planner/logical_operator_visitor.cpp
… src/planner/operator/logical_insert.cpp
… src/storage/serialization/serialize_logical_operator.cpp
… test/sql/generated_columns/virtual/update_index.test
… test/sql/index/art/constraints/test_art_tx_upserts_local.test
… test/sql/upsert/postgres/non_spurious_duplicate_violation.test
… test/sql/upsert/test_big_insert_no_vector_verification.test
… test/sql/upsert/upsert_aliased.test
… test/sql/upsert/upsert_basic.test
… test/sql/upsert/upsert_default_expressions.test
… test/sql/upsert/upsert_excluded_references.test
… test/sql/upsert/upsert_global_mixed_conflicts.test
… tools/pythonpkg/requirements-dev.txt