https://github.com/duckdb/duckdb/pull/18135 introduced the `MERGE INTO` operator.
Merge allows us to do upserts - similar to `ON CONFLICT`, but without requiring the target table to have an index.

The merge operator has a separate implementation that leans on the existing insert/update/delete operators.
The `ON CONFLICT` instead has a fully custom implementation that inlines upserting into the `PhysicalInsert`.

This has a number of issues:
* The `ON CONFLICT` code is complex, and we now have an alternative code path that can be used to achieve the same functionality. Maintaining both is sub-optimal.
* The `ON CONFLICT` code always uses the index for detecting conflicts - effectively forcing an index join to be used

This PR unifies the implementation of these two by rewriting the `ON CONFLICT` clause in the binder to an identical `MERGE` statement, i.e. this:

```sql
INSERT OR REPLACE INTO tbl FROM source;
```

Becomes (under the hood):

```sql
MERGE INTO tbl USING (SELECT DISTINCT ON (key) * FROM source) ON (tbl.key = source.key)
WHEN MATCHED THEN UPDATE
WHEN NOT MATCHED THEN INSERT
```

### Behavioral Changes

#### Multi-Row Updates

The previous upsert implementation threw an error if the same row was updated multiple times in the same statement. This functionality has been removed - this is now allowed. 

#### WHERE clause changes

The ON CONFLICT clause in Postgres has the ability to specify a WHERE clause after the indexed columns as part of the conflict target, e.g.:

```sql
ON CONFLICT (a) WHERE index_predicate [conflict_action]
```

This predicate is used to refer to partial indexes. 

> index_predicate
> Used to allow inference of partial unique indexes. Any indexes that satisfy the predicate (which need not actually be partial indexes) can be inferred. Follows CREATE INDEX format. SELECT privilege on any column appearing within index_predicate is required.

DuckDB does not support partial indexes. However, this WHERE clause was instead mistakenly used as a WHERE clause on the condition.

There is a second WHERE clause that can be specified as part of DO UPDATE SET that is instead used for this:

```sql
    DO UPDATE SET ...
              [ WHERE condition ]
```
 
After this PR, we now throw an error if the WHERE clause is used as part of the conflict target as we don't support partial indexes - and instead only support WHERE as part of DO UPDATE SET.

#### WHERE fallback

In the previous implementation, if the WHERE clause was not met, we would fallback to inserting. This is not particularly useful, as this would always result in a constraint violation. After all: the `CONFLICT` clause matched for that particular row. This also does not match Postgres' behavior - they fall back to `DO NOTHING` (i.e. ignoring the row). We now implement the same fallback as Postgres.

We also no longer support WHERE as part of `DO NOTHING`, as that does not have any purpose anymore (the fallback would also be `DO NOTHING`).

### (De)serialization

As we do the rewrite in the binder, this does not apply when deserializing plans from older versions. As such in order to support deserializing older plans, for now the old code is still kept in. It should just not be triggered anymore except when deserializing old plans.

### Benchmarks

The `MERGE` code-path automatically switches between an index join (using pushdown) and a full hash join - allowing it to be much faster for bulk appends. Below is a benchmark running `INSERT OR REPLACE` on a full TPC-H SF1 lineitem table with primary key constraint defined:

```sql
INSERT OR REPLACE INTO target_lineitem FROM lineitem
```

| main  | New  |
|-------|------|
| 23.2s | 2.3s |